+++ CHANGELOG.md
@@ -3,7 +3,15 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+* Flag: --log-level-switcher-listen-addr (default:1065) to change log level on a running instance (see DEBUG.md)
+* Flag: --pprof-listen-addr (default: 6060)
+
+
 ## [v0.1.0-beta3] 2020-05-13
+* Added command `kv` to `tools` with sub command `get`, `scan`, `prefix`, `account`, `blk`, `blkirr`, `trx`, `trxtrace` to retrieve data from trxdb 
 * Added `--eosq-available-networks` json string to configure the network section of eosq. 
 ``` [
      {